pyproject.toml
[tool.poetry]
name = "viso-jax"
version = "0.1.0"
description = "GPU-accelerated value iteration and simulation optimization for perishable inventory control using JAX"
authors = ["Joe Farrington <farrington.joe@gmail.com>"]
readme = "README.md"
packages = [{include = "viso_jax"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
hydra-core = "^1.2.0"
numpy = "^1.23.3"
pandas = "^1.5.0"
jaxlib = {version = "^0.3.15+cuda11.cudnn82", source = "jaxlib"}
jax = {extras = ["cuda"], version = "^0.3.15", source = "jax"}
matplotlib = "^3.6.2"
numpyro = "^0.10.1"
gymnax = "^0.0.5"
tensorflow-probability = "^0.19.0"
optuna = "^3.0.5"
plotly = "^5.11.0"
seaborn = "^0.12.1"

[[tool.poetry.source]]
name = "jax"
url = "https://storage.googleapis.com/jax-releases/jax_releases.html"
default = false
secondary = false

[[tool.poetry.source]]
name = "jaxlib"
url = "https://storage.googleapis.com/jax-releases/jax_cuda_releases.html"
default = false
secondary = false

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
black = "^22.12.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.1.3"
pytest-datadir = "^1.4.1"

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "serial",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"